backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
firebase-admin==6.5.0
google-generativeai==0.7.2
PyMuPDF==1.24.10
python-dotenv==1.0.1
reportlab==4.2.2
gunicorn==22.0.0
Werkzeug==3.0.3
requests==2.32.3
